delegation is clearer than interface
